Created Function that Validates Client Quantity
The function that validates if the input for clients quantity is empty, if the quantity of clientsexceeds the booking segment capacity and if there are invalid characters in the input has been declared. The Unitary Tests to see if the function is working correctly have also been added on the unitaryTests.js file.