Validating Functions and Unitary Tests added
The functions that validate the booking hour and the client cellphone were added to the main JS file. The client cliencellphone variable was changed to be declared when the submit button is interacted with. The unitary test that validates the booking hour has also been added in the unitaryTests JS file.