Restructured main.js file
The main.js file has been restructured for better readability.
As the validate function is the first to run to check if everything that has been inserted into the input files is correctly, it should be on top.

The error handling function has also been updated to return a counter which if higher than zero, it wont let run the functions that add a new booking.