Added if statement and modified CheckAvailableSpace function
An "If" statement was added to the validateForm function that blocks the submission if there's no available space in the booking hour segment.

The if statements that should happen when the space is unavailable were added. They now show an error when there's no space available to create the booking.

A typo has also been corrected from the main.js file and the unitaryTests.js file.