Created Function that Deletes Previous Errors
The function deletePreviousErrors runs everytime the form is trying to be validated. It basically gets every element inside an array and then it deletes them run by a callback function that is inside a forEach function. This function was needed so the elements that remained from the last validation don't overstay their welcome.