Created Function that Deletes Existing Bookings
The basic function that deletes pre-existing bookings has been created.
This was created to solve the problem arises when an existing booking is canceled by a client.
This was achieved by assiging a checkbox to each booking that its created, then each checkbox is searched by a document query and if they result checked, they will be deleted